+++ piecrust-uplink/CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `emit_raw` and `feed_raw` functions
 - Add `ContractError::DoesNotExist` variant
 
+### Changed
+
+- Remove `'static` bound on `emit` topics
+
 ## [0.16.0] - 2024-08-01
 
 ### Added